+++ src/main/java/me/recipe/homework/service/RecipeService.java
@@ -2,14 +2,11 @@ package me.recipe.homework.service;
 
 import me.recipe.homework.model.Recipe;
 
-import java.io.IOException;
+import java.io.File;
 import java.nio.file.Path;
 import java.util.Collection;
 
 public interface RecipeService {
-
-    static void saveToFile(String json) {
-    }
 
     static String readFromFile() {
         return readFromFile();
@@ -31,6 +28,15 @@ public interface RecipeService {
     void deleteAllRecipe();
 
     Path createRecipesFile();
+
+    File getDataFile();
+
+    boolean cleanDateFile();
+
+    static void saveToFile(String json) {
+    }
+
+    boolean saveToFile(String json);
 
     byte[] exportTxt();
 }
